% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{a.spambase.analysis}
\alias{a.spambase.analysis}
\title{Main function which guides through the data analysis.}
\usage{
a.spambase.analysis()
}
\description{
Main function which guides through the data analysis.
}

