% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory.R
\name{b.b.exploratory.structure_data}
\alias{b.b.exploratory.structure_data}
\title{Structure the spambase data}
\usage{
b.b.exploratory.structure_data()
}
\description{
To get (back) to the overview of all steps and functions use this link:
\code{\link{a.a.main}}

It creates an environment variable \code{e} which stores all types types in its own variable.
}
\author{
Vitali Friesen
}

