% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory.R
\name{b.c.exploratory.str_summary}
\alias{b.c.exploratory.str_summary}
\title{BBBB}
\usage{
b.c.exploratory.str_summary()
}
\description{
To get (back) to the overview of all steps and functions use this link: \code{\link{a.a.main}}
}
\author{
Vitali Friesen
}

