% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory.R
\name{b.exploratory.structure_data}
\alias{b.exploratory.structure_data}
\title{Structure the spambase data}
\usage{
b.exploratory.structure_data()
}
\description{
Create an environment variable \code{stuff} which
}

