% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality.R
\name{c.a.normality}
\alias{c.a.normality}
\title{C}
\usage{
\method{c}{a.normality}()
}
\description{
To get (back) to the overview of all steps and functions use this link: \code{\link{a.a.main}}
}
\author{
Vitali Friesen
}

