% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality.R
\name{c.c.chisquare}
\alias{c.c.chisquare}
\title{CCC}
\usage{
\method{c}{c.chisquare}()
}
\description{
To get (back) to the overview of all steps and functions use this link: \code{\link{a.a.main}}
}
\author{
Vitali Friesen
}

