% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory.R
\name{exploratory.data_analysis}
\alias{exploratory.data_analysis}
\title{Exploratory data analysis wrapper function}
\usage{
exploratory.data_analysis()
}
\description{
This is a wrapper function for the whole exploratory data analysis step
}

