% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{z.c.boxcox}
\alias{z.c.boxcox}
\title{A function for transforming data with boxcox}
\usage{
z.c.boxcox(data, lambda)
}
\arguments{
\item{data}{The matrix or data.frame which gets transformed.}

\item{lambda}{The intensity of the transformation.}
}
\value{
The transformed data
}
\description{
To get (back) to the overview of all steps and functions use this link: \code{\link{a.a.main}}

This function transforms data with the boxcox transformation to reduce scews.
}
\author{
Vitali Friesen (Credits: Tutorial of Data Analytics 2016 by PROF. DR. HEIKE TRAUTMANN )
}

