% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{z.d.loglik}
\alias{z.d.loglik}
\title{Likelihood function for boxcox transformation}
\usage{
z.d.loglik(lambda, data)
}
\description{
To get (back) to the overview of all steps and functions use this link: \code{\link{a.a.main}}

Likelihood function to calculate how close the boxcox transformation is to being normal.

Can be used to be optimized to find optimal lambda.
}
\author{
Vitali Friesen (Credits: Tutorial of Data Analytics 2016 by PROF. DR. HEIKE TRAUTMANN )
}

