% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{z.z.util.test.func}
\alias{z.z.util.test.func}
\title{Example Function for docu creation}
\usage{
z.z.util.test.func(x, y)
}
\arguments{
\item{x}{A number which will be squared}

\item{y}{A number which get multiplied to \code{x^2}}
}
\value{
the caculated result of \code{x * x * y}
}
\description{
Example function for understanding how to document
Calculates product of \code{x^2} multiplied by \code{y}
}
\examples{
test.func(3,2)
# 18

test.func(2,3)
# 12
}

