% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory.R
\name{b.a.exploratory.data_analysis}
\alias{b.a.exploratory.data_analysis}
\title{Exploratory - Data Analysis Wrapper}
\usage{
b.a.exploratory.data_analysis()
}
\description{
To get (back) to the overview of all steps and functions use this link:
\code{\link{a.a.main}}

This is a wrapper function for the whole exploratory data analysis step.

First the data is stored in different variables which will then be accessable from the
environment.

Then different visualizations are created which are summaries, correlations and dimension
visualization.

It executes four functions:
\enumerate{
  \item \code{\link{b.b.exploratory.structure_data}}
  \item \code{\link{b.c.exploratory.str_summary}}
  \item \code{\link{b.d.exploratory.correlations}}
  \item \code{\link{b.e.exploratory.dim_vis}}
}
}
\author{
Vitali Friesen
}

