% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory.R
\name{b.a.exploratory.data_analysis}
\alias{b.a.exploratory.data_analysis}
\title{Exploratory data analysis wrapper function}
\usage{
b.a.exploratory.data_analysis()
}
\description{
This is a wrapper function for the whole exploratory data analysis step
}

