% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory.R
\name{b.d.exploratory.correlations}
\alias{b.d.exploratory.correlations}
\title{BBBBB}
\usage{
b.d.exploratory.correlations()
}
\description{
To get (back) to the overview of all steps and functions use this link: \code{\link{a.a.main}}
}
\author{
Vitali Friesen
}

