% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality.R
\name{c.d.distribution_test}
\alias{c.d.distribution_test}
\title{CCCC}
\usage{
\method{c}{d.distribution_test}()
}
\description{
To get (back) to the overview of all steps and functions use this link: \code{\link{a.a.main}}
}
\author{
Vitali Friesen
}

