% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality.R
\name{c.e.norm_transform}
\alias{c.e.norm_transform}
\title{CCCCC}
\usage{
\method{c}{e.norm_transform}()
}
\description{
To get (back) to the overview of all steps and functions use this link: \code{\link{a.a.main}}
}
\author{
Vitali Friesen
}

